reports: Store by default json report and test plan
This aligns reporting of V2 with V1
