+++ tests/conftest.py
@@ -1,3 +1,4 @@
+from collections import namedtuple
 from pathlib import Path
 from unittest import mock
 
@@ -48,4 +49,17 @@ def mock_get_toolchain_version():
     """
     with mock.patch('twister2.environment.environment._get_toolchain_version_from_cmake_script') as mocked_object:
         mocked_object.return_value = 'zephyr'
+        yield mocked_object
+
+
+@pytest.fixture(scope='function', autouse=True)
+def mock_get_zephyr_repo_info():
+    """
+    we need to mock
+    twister2.environment.environment.get_zephyr_repo_info
+    because in unit tests we don't have Zephyr repo to call this script
+    """
+    with mock.patch('twister2.environment.environment.get_zephyr_repo_info') as mocked_object:
+        RepoInfo = namedtuple('RepoInfo', 'zephyr_version commit_date')
+        mocked_object.return_value = RepoInfo('123456789012', '20220102')
         yield mocked_object
